From Micheal Witten.
I thought it would be pretty if the alternately colored
boxes were `ordered' along the visible spectrum:

  1 touch   : red
  2 touches : green
  3 touches : blue

In this way, the player's ship drains the box of energy
until the box is no more  :-P 

Also, the red goes quite well with other red aspects of
the playing area (the HUD).

To be pedantic, the code has also been refactored to
reflect this ordering as much as possible.