"Crazy" -> "difficult"
The use of the term "crazy" to refer to things with negative connotations
is a form of casual sanism that unfortunately populates the English lexicon.
The term "difficult" does not carry these connotations and is more precise,
so this replaces the term "crazy" with "difficult" in the game's help.

"Crazy Eights" was left alone here because it's a reference to a card
game. That game's name itself does have sanist origins, referring to
mentally ill soldiers that were discharged from the military, but
unfortunately the game isn't known to anyone by its original name
"Eights" and the sanist origin of the game's name is inobvious.
Perhaps changing the name of the "Crazy Eights" level to just "Eights",
or possibly an alternative like "Uno", may be justified.